Update code to use single instance of `ContentAdapter`
In light of the desired interaction between `ContentAdapter` and the View, the
implementation was updated to use only a single instance of the adapter as a
means to allow proper inversion of control. In doing such, part of the logic
within the adapter needed to be updated to better reflect such change.